save: left only migrated file; finished migrating
