--- MIGRATED/src/main.ts
@@ -1,5 +1,0 @@
-import { createApp } from 'vue';
-import './style.css';
-import App from './App.vue';
-
-createApp(App).mount('#app');
