--- STASH/.gitignore
@@ -1,38 +1,0 @@
-logs
-*.log
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-lerna-debug.log*
-report.[0-9]*.[0-9]*.[0-9]*.[0-9]*.json
-pids
-*.pid
-*.seed
-*.pid.lock
-.DS_Store
-lib-cov
-coverage
-*.lcov
-.nyc_output
-.lock-wscript
-build/Release
-node_modules/
-jspm_packages/
-typings/
-*.tsbuildinfo
-.npm
-.eslintcache
-.node_repl_history
-*.tgz
-.yarn-integrity
-.env
-.env.test
-.cache
-.next
-.nuxt
-.vuepress/dist
-.serverless/
-.fusebox/
-.dynamodb/
-.webpack/
-out/
