--- STASH/.npmrc
@@ -1,1 +1,0 @@
-save-exact = true
