--- STASH/.vscode/extensions.json
@@ -1,3 +1,0 @@
-{
-  "recommendations": ["Vue.volar", "Vue.vscode-typescript-vue-plugin"]
-}
