--- STASH/.vscode/settings.json
@@ -1,5 +1,0 @@
-{
-    "search.exclude": {
-        "**/node_modules": true
-    }
-}
