--- STASH/README.md
@@ -1,35 +1,0 @@
-# Wiki Search
-
-## Getting Started
-
-#### Prerequisites
-
-###### Node JS
-
-You need to have Node JS installed on your machine, as well as the npm package manager.
-To download both, go [here](https://nodejs.org/en/download) and download Node js.
-
-#### Installation
-
-1. Clone the repo
-```
-git clone https://github.com/esccaa/Wiki-Search.git
-```
-2. Install Electron js module
-```
-npm install electron --save-dev
-```
-3. To start the program
-```
-npm start
-```
-
-## Usage
-
-## Roadmap
-
-## Contributing
-
-## License
-
-## Ackknowledgements
