--- STASH/package.json
@@ -1,56 +1,0 @@
-{
-  "name": "wiki-search",
-  "productName": "",
-  "version": "1.0.0",
-  "description": "A computer application for searching Wikipedia articles",
-  "type": "module",
-  "main": "src/index.ts",
-  "scripts": {
-    "start": "electron-forge start",
-    "package": "electron-forge package",
-    "make": "electron-forge make",
-    "publish": "electron-forge publish",
-    "lint": "echo \"No linting configured\""
-  },
-  "keywords": [],
-  "author": "Geonwoo Kim",
-  "license": "MIT",
-  "dependencies": {
-    "@vitejs/plugin-vue": "4.1.0",
-    "csstype": "3.1.2",
-    "electron-squirrel-startup": "^1.0.0",
-    "esbuild": "0.17.17",
-    "estree-walker": "3.0.3",
-    "fsevents": "2.3.2",
-    "function-bind": "1.1.1",
-    "has": "1.0.3",
-    "is-core-module": "2.12.0",
-    "magic-string": "0.30.0",
-    "nanoid": "4.0.2",
-    "path-parse": "1.0.7",
-    "resolve": "1.22.3",
-    "rollup": "3.20.3",
-    "source-map": "0.7.4",
-    "source-map-js": "1.0.2",
-    "sourcemap-codec": "1.4.8",
-    "supports-preserve-symlinks-flag": "1.0.0",
-    "vite": "4.2.1",
-    "vu": "0.0.0",
-    "vue": "3.2.47",
-    "xml2js": "0.5.0"
-  },
-  "devDependencies": {
-    "@electron-forge/cli": "^6.0.5",
-    "@electron-forge/maker-deb": "^6.0.5",
-    "@electron-forge/maker-rpm": "^6.0.5",
-    "@electron-forge/maker-squirrel": "^6.0.5",
-    "@electron-forge/maker-zip": "^6.0.5",
-    "@vitejs/plugin-vue": "^4.0.0",
-    "electron": "^23.1.3",
-    "typescript": "5.0.4",
-    "vite": "^4.1.4"
-  },
-  "directories": {
-    "doc": "docs"
-  }
-}
