--- STASH/src/front/liked.vue
@@ -1,13 +1,0 @@
-<template>
-
-</template>
-
-
-<script>
-
-</script>
-
-
-<style>
-
-</style>
