+++ src/front/settings.html
